+++ remote-helloworld-mdb/README.adoc
@@ -55,12 +55,6 @@ $ docker run --rm --name artemis -e AMQ_USER=admin -e AMQ_PASSWORD=admin -p8161:
 
 You can access the console of {BrokerProductName} at the URL: http://localhost:8161/console with *admin/admin* to authenticate.
 
-// Back Up the {productName} Standalone Server Configuration
-include::../shared-doc/back-up-server-standalone-configuration.adoc[leveloffset=+2]
-
-// Start the {productName} Standalone Server
-include::../shared-doc/start-the-standalone-server.adoc[leveloffset=+2]
-
 [[configure_the_server]]
 === Configure the {productName} Server
 
@@ -103,8 +97,8 @@ The following `testQueue` jms-queue was configured in the configuration of the `
 <pooled-connection-factory name="RemoteConnectionFactory" entries="java:jboss/RemoteConnectionFactory java:jboss/exported/jms/RemoteConnectionFactory" connectors="artemis" user="admin" password="admin" enable-amq1-prefix="false"/>
 ----
 
-
-
+// Start the {productName} Standalone Server
+include::../shared-doc/start-the-standalone-server.adoc[leveloffset=+2]
 // Build and Deploy the Quickstart
 include::../shared-doc/build-and-deploy-the-quickstart.adoc[leveloffset=+2]
 
@@ -125,12 +119,8 @@ INFO  [class org.jboss.as.quickstarts.mdb.HelloWorldQueueMDB] (Thread-5 (ActiveM
 INFO  [class org.jboss.as.quickstarts.mdb.HelloWorldQueueMDB] (Thread-4 (ActiveMQ-client-global-threads-1189700957)) Received Message from queue: This is message 5
 ----
 
-// Testing with Arquillian
-include::../shared-doc/run-arquillian-integration-tests-with-server-distribution.adoc[leveloffset=+2]
-// Undeploy the Quickstart
-include::../shared-doc/undeploy-the-quickstart.adoc[leveloffset=+2]
-// Restore the {productName} Standalone Server Configuration
-include::../shared-doc/restore-standalone-server-configuration.adoc[leveloffset=+2]
+// Server Distribution Testing
+include::../shared-doc/run-integration-tests-with-server-distribution.adoc[leveloffset=+2]
 
 // Additional information about this script
 This script removes the remote broker connection from the `messaging-activemq` subsystem and restore the `default` internal broker. You should see the following result when you run the script:
@@ -147,18 +137,19 @@ This script removes the remote broker connection from the `messaging-activemq` s
 // Restore the {productName} Standalone Server Configuration Manually
 include::../shared-doc/restore-standalone-server-configuration-manual.adoc[leveloffset=+3]
 
+include::../shared-doc/undeploy-the-quickstart.adoc[leveloffset=+2]
+
+[NOTE]
+====
+The integration tests expect a running {BrokerProductName} broker, so make sure you have started the broker before you begin.
+====
 // Build and run sections for other environments/builds
 ifndef::ProductRelease,EAPXPRelease[]
 include::../shared-doc/build-and-run-the-quickstart-with-provisioned-server.adoc[leveloffset=+1]
-[NOTE]
-====
-The Arquillian integration tests expect a running {BrokerProductName} broker, so make sure you have started the broker before you begin.
-====
 
 endif::[]
 
-== Building and running the quickstart application with OpenShift
-
+include::../shared-doc/build-and-run-the-quickstart-with-openshift.adoc[leveloffset=+1]
 === Deploy a {BrokerProductName} instance on OpenShift
 
 [source,options="nowrap",subs="+attributes"]
@@ -169,12 +160,7 @@ $ oc run artemis --env AMQ_USER=admin --env AMQ_PASSWORD=admin --image={BrokerIm
 This will create a broker instance named `artemis` on OpenShift that can be accessed on the port `61616` on the service `artemis` with the account *admin/admin*.
 //Prepare Helm for Quickstart Deployment
 
-// Getting Started with Helm
-include::../shared-doc/helm-getting-started-overview.adoc[leveloffset=+2]
-//Prepare Helm for Quickstart Deployment
-include::../shared-doc/helm-deploy-project.adoc[leveloffset=+2]
-// Testing on Openshift
-include::../shared-doc/run-arquillian-integration-tests-with-openshift.adoc[leveloffset=+2]
+include::../shared-doc/build-and-run-the-quickstart-with-openshift.adoc[leveloffset=+1]
 
 ==== Clean Up
