--- remote-helloworld-mdb/src/test/java/org/jboss/as/quickstarts/mdb/RemoteMDBServletIT.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2022 JBoss by Red Hat.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.as.quickstarts.mdb;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
-import org.junit.runner.RunWith;
-
-/**
- *
- * @author Emmanuel Hugonnet (c) 2023 Red Hat, Inc.
- */
-@RunWith(Arquillian.class)
-@RunAsClient
-public class RemoteMDBServletIT extends AbstractMDBServletIT {
-
-
-    private String getServerHost() {
-        String host = System.getenv("SERVER_HOST");
-        if (host == null) {
-            host = System.getProperty("server.host", "http://localhost:8080/remote-helloworld-mdb");
-        }
-        return host;
-    }
-
-    @Override
-    protected URI getHTTPEndpoint() {
-        try {
-            return new URI(getServerHost() + "/HelloWorldMDBServletClient");
-        } catch (URISyntaxException ex) {
-            throw new RuntimeException(ex);
-        }
-    }
-}
